+++ tests/fixtures/noqa/noqa.py
@@ -703,3 +703,9 @@ for element in range(10):
         # https://github.com/wemake-services/wemake-python-styleguide/issues/1082
         break
     my_print(4)
+
+
+class Mem0Output(object):  # noqa: WPS124
+    # See:
+    # https://github.com/wemake-services/wemake-python-styleguide/issues/1191
+    anti_wps124 = 'unreadable class'