+++ wemake_python_styleguide/visitors/ast/naming.py
@@ -20,13 +20,15 @@ from wemake_python_styleguide.compat.types import AnyAssignWithWalrus
 from wemake_python_styleguide.constants import (
     MODULE_METADATA_VARIABLES_BLACKLIST,
     SPECIAL_ARGUMENT_NAMES_WHITELIST,
+    UNREADABLE_CHARACTER_COMBINATIONS,
 )
 from wemake_python_styleguide.logic import nodes
 from wemake_python_styleguide.logic.naming import (
     access,
+    alphabet,
     blacklists,
     builtins,
-    logical,
+    name_check,
     name_nodes,
 )
 from wemake_python_styleguide.logic.tree import functions
@@ -72,19 +74,27 @@ class _NameValidator(object):
         *,
         is_first_argument: bool = False,
     ) -> None:
-        if logical.is_wrong_name(name, self._variable_names_blacklist):
+        if name_check.is_wrong_name(name, self._variable_names_blacklist):
             self._error_callback(
                 naming.WrongVariableNameViolation(node, text=name),
             )
 
         if not is_first_argument:
-            if logical.is_wrong_name(name, SPECIAL_ARGUMENT_NAMES_WHITELIST):
+            if name_check.is_wrong_name(name, SPECIAL_ARGUMENT_NAMES_WHITELIST):
                 self._error_callback(
                     naming.ReservedArgumentNameViolation(node, text=name),
                 )
 
-        if logical.does_contain_unicode(name):
+        if alphabet.does_contain_unicode(name):
             self._error_callback(naming.UnicodeNameViolation(node, text=name))
+
+        unreadable_sequence = alphabet.get_unreadable_characters(
+            name, UNREADABLE_CHARACTER_COMBINATIONS,
+        )
+        if unreadable_sequence:
+            self._error_callback(
+                naming.UnreadableNameViolation(node, text=unreadable_sequence),
+            )
 
         self._ensure_length(node, name)
         self._ensure_underscores(node, name)
@@ -117,12 +127,12 @@ class _NameValidator(object):
                 naming.PrivateNameViolation(node, text=name),
             )
 
-        if logical.does_contain_underscored_number(name):
+        if alphabet.does_contain_underscored_number(name):
             self._error_callback(
                 naming.UnderscoredNumberNameViolation(node, text=name),
             )
 
-        if logical.does_contain_consecutive_underscores(name):
+        if alphabet.does_contain_consecutive_underscores(name):
             self._error_callback(
                 naming.ConsecutiveUnderscoresInNameViolation(
                     node, text=name,
@@ -141,18 +151,26 @@ class _NameValidator(object):
 
     def _ensure_length(self, node: ast.AST, name: str) -> None:
         min_length = self._options.min_name_length
-        if logical.is_too_short_name(name, min_length=min_length):
-            self._error_callback(naming.TooShortNameViolation(node, text=name))
+        if name_check.is_too_short_name(name, min_length=min_length):
+            self._error_callback(
+                naming.TooShortNameViolation(
+                    node, text=name, baseline=min_length,
+                ),
+            )
 
         max_length = self._options.max_name_length
-        if logical.is_too_long_name(name, max_length=max_length):
-            self._error_callback(naming.TooLongNameViolation(node, text=name))
+        if name_check.is_too_long_name(name, max_length=max_length):
+            self._error_callback(
+                naming.TooLongNameViolation(
+                    node, text=name, baseline=max_length,
+                ),
+            )
 
     def _ensure_case(self, target: ast.AST) -> None:
         if not isinstance(target, ast.Name):
             return
 
-        if not target.id or not logical.is_upper_case_name(target.id):
+        if not target.id or not name_check.is_upper_case_name(target.id):
             return
 
         self._error_callback(
@@ -190,6 +208,7 @@ class WrongNameVisitor(BaseNodeVisitor):
             UpperCaseAttributeViolation
             UnicodeNameViolation
             TrailingUnderscoreViolation
+            UnreadableNameViolation
 
         """
         self._validator.check_attribute_name(node)
@@ -207,6 +226,7 @@ class WrongNameVisitor(BaseNodeVisitor):
             TooLongNameViolation
             UnicodeNameViolation
             TrailingUnderscoreViolation
+            UnreadableNameViolation
 
         """
         self._validator.check_name(node, node.name)
@@ -238,6 +258,7 @@ class WrongNameVisitor(BaseNodeVisitor):
             PrivateNameViolation
             TooLongNameViolation
             TrailingUnderscoreViolation
+            UnreadableNameViolation
 
         """
         for alias_node in node.names:
@@ -257,6 +278,7 @@ class WrongNameVisitor(BaseNodeVisitor):
             TooLongNameViolation
             UnicodeNameViolation
             TrailingUnderscoreViolation
+            UnreadableNameViolation
 
         """
         variable_name = name_nodes.get_assigned_name(node)