+++ CHANGELOG.md
@@ -47,6 +47,8 @@ Semantic versioning in our case means:
 - Adds baseline information for all complexity violation messages: `x > baseline`
 - Changes how cognitive complexity is calculated
 - Adds support for positional arguments in different checks
+- Adds `UnreadableNameViolation` as `WPS124` because there are some
+character combination which is not easy to read
 - Adds support for `NamedExpr` with in compare type violation
 
 ### Bugfixes